extract mappings
